requirements.txt
flask~=3.1.1
python-dotenv~=1.1.0
psycopg2-binary~=2.9.10
pandas~=2.2.3
numpy~=2.0.2
sqlalchemy~=2.0.41
pip~=25.1
wheel~=0.41.2
pytz~=2025.2
MarkupSafe~=3.0.2
werkzeug~=3.1.3
click~=8.1.8
jinja2~=3.1.6
blinker~=1.9.0
itsdangerous~=2.2.0
setuptools~=78.1.1
python-dateutil~=2.9.0.post0
six~=1.17.0
pytest~=8.4.1
gunicorn==21.2.0